==> Jupyter_Notebooks/2_ETL_User_Items.ipynb <==
### **ETL process (Extract, Transform & Load) - Plataforma de videojuegos Stream** 
#### **DATOS: User_Items**

Empezaremos por entender, transfomar y disponibilizar, encontrando problemas en los datos, aplicando tecnicas de limpieza y preprocesamiente. Finalmente, almacenaremos los datos transformatos para futuras exploraciones.  

Entenderemos el contexto y la informacion necesaria para nuestro analisis, tratando consistencia a esos datos.

#### **Importamos las librerias necesarias**

import numpy as np # Para trabajar con arrays y matrices


import pandas as pd #para el analisis de tablas tabulares
from pandas import json_normalize 

import os # OS module help us interact with the operative sistem

# Gdown makes easy to download files from Google Drive with its ID 
import gdown

# Json its a commonly used format for exchange of data
import json


import warnings
warnings.filterwarnings("ignore")

import ast

### 2º manera de subir el raw: ESTO es para subir desde un gzip raw

# import gzip 
# import ast
# import pandas as pd
# data = []
# try:
#     with gzip.open('users_items.json.gz', 'rb') as d:
#         for line in d:
#             data.append (ast.literal_eval(line.decode('utf-8')))
#     Users_Items = pd.DataFrame (data)
# finally:
#     d.close()

## **1. Cargamos el Dataset**
Fuente de datos: **australian_user_items.json**

Vamos a descargar el archivo de forma local, ya que es necesario cargarlo de una manera especifica, debido a su estructura. 

Este archivo está disponible en Google Drive con acceso compartido. Puede descargarlo en el siguiente enlace:

Datasets: https://drive.google.com/uc?export=download&id=1BzbeJv1878pa2lQMspr5y03dUdIXaYyX


ruta = '/Users/cristian/Desktop/ProyectoSteam/Jupyter_Notebooks/user_items.json'

with open(ruta, 'r', encoding='utf-8') as f:
    data = f.readlines()

# vamos a convertir las lineas en registros
records = [eval(line.strip()) for line in data]

# Creamos el DataFrame a partir de los records o registros
df_UserItems= pd.DataFrame(records)

respaldo = df_UserItems

## **2. Explorar y entender el conjunto de datos**

# traemos una vista rapida de la tabla

print('DataFrame - First & Last Rows:')
df_UserItems

### **# Analisis Primario**
1. El dataset 'User_items' no muestra datos nulos en ninguna columna.

2. El proceso de recoleccion de datos esta estructurado de una manera optimizada ya que no hay ningun valor faltante. 

3. La columna 'items' tiene una lista anidada

4. Se realizara hacer un analisis mas profundo. 

# Vamos a conocer cuantas Filas tienen nulos en todas sus columnas

print(f'total filas de games_df: {df_UserItems.shape[0]}')
print(f'total filas de games_df con nulos en todas sus columnas: {df_UserItems.isnull().all(axis=1).sum()}')


# Obtener información general del DataFrame
print("Información general del DataFrame:\n")
df_UserItems.info()

### **# Insights**
1. La columna **'items'** contiene una lista con datos anidados (lista de diccionarios).

2. Se va a normalizar la columna. Con esto vamos a obtener una columna por cada clave en ese diccionario, mandentremos la trazabilidad de las columnas **('steam_id', 'items_count', 'user_id' y 'user_url')**
   



# Normalizar la columna 'items'
df_UserItems = pd.json_normalize(records, record_path=['items'], meta=['steam_id','items_count','user_id','user_url'] )

df_UserItems

# buscamos los registros duplicados
duplicados = df_UserItems[df_UserItems.duplicated(subset=['steam_id','item_id','playtime_forever'])]
duplicados.shape

#Borramos dulplicados
df_UserItems = df_UserItems.drop_duplicates(keep='first')

# Convertir la columna 'playtime_forever' de minutos a horas
df_UserItems['hours_game'] = df_UserItems['playtime_forever'] / 60

# Redondear los valores a dos decimales (opcional)
df_UserItems['hours_game'] = df_UserItems['hours_game'].round(2)

# Obtener los 5 valores máximos
top_5_max = df_UserItems.nlargest(5, 'hours_game')

# Obtener los 5 valores mínimos
top_5_min = df_UserItems.nsmallest(5, 'hours_game')

# Mostrar los resultados
print("Top 5 valores máximos:")
print(top_5_max[['hours_game']])

print("\nTop 5 valores mínimos:")
print(top_5_min[['hours_game']])

# Contar las filas con valor 0 en la columna 'hours_game'
filas_con_cero = (df_UserItems['hours_game'] == 0).sum()

# Imprimir el resultado
print(f"Existen {filas_con_cero} filas con valor 0 en la columna 'hours_game'.")

# Eliminar las filas con valor 0 en la columna 'hours_game'
df_UserItems = df_UserItems[df_UserItems['hours_game'] != 0]

#Eliminamos columna no relevantes
df_UserItems = df_UserItems.drop(['playtime_2weeks','user_url','items_count','item_name','steam_id','playtime_forever'], axis=1)
df_UserItems.columns

### **# Insights**
1. Se eliminan los registros con valor nulo en esta columna por motivos de optimizacion de consulta.
   
2. Se encuentran 1'847,730 registros con un valor de 0 en la columna **'hours_game'**.

3. Hay columas irrelevantes, que tambien se remueven por motivos de optimizacion de consultas. 
   



df_UserItems

df_UserItems.info()

### 4. Guardamos los datos limpios
Guardamos estos archivos en los siguientes formatos para disponibilizarlos. CSV/ Json/ Parquet

# Los archivos se almacenan en local 
# df_UserItems.to_csv('user_items_limpios.csv', index=False)
# df_UserItems.to_json('user_items_limpios.json', orient='records', lines=True)
df_UserItems.to_parquet('../Dataframes/Dataframes_limpios/user_items_limpios.parquet', index=False)
