==> Jupyter_Notebooks/2_User_Items_ETL_.ipynb <==
### ETL Process (Extract, Transform & Load) - Stream Video Games Platform
#### Data: User_Items

We will start by understanding, transforming, and making the data available by identifying issues, applying cleaning techniques, and preprocessing. Finally, we will store the transformed data for future explorations.

We will understand the context and the necessary information for our analysis, ensuring data consistency.

#### Importing necessary libraries:

- `numpy`: For working with arrays and matrices.
- `pandas`: For tabular data analysis.
- `os`: Allows interaction with the operating system.
- `gdown`: Facilitates downloading files from Google Drive.
- `json`: Commonly used for data exchange.
- `warnings`: Filters and suppresses warnings.
- `ast`: Useful for parsing Python code at a deeper level than string parsing.


import numpy as np  # For array and matrix operations
import pandas as pd  # For tabular data analysis
from pandas import json_normalize  # For normalizing JSON data

import os  # For interacting with the operating system

import gdown  # For downloading files from Google Drive

import json  # Common data exchange format

import warnings
warnings.filterwarnings("ignore")  # Suppresses warnings

import ast  # For parsing Python code



### **1. Load Dataset**

Data Source: australian_user_items.json

We'll download the file locally since it needs to be loaded in a specific way due to its structure.

This file is available on Google Drive with shared access. You can download it from the following link:


Dataset: https://drive.google.com/uc?export=download&id=1BzbeJv1878pa2lQMspr5y03dUdIXaYyX


path= '../Data_Files/raw_Data/users_items.json'


with open(path, 'r', encoding='utf-8') as f:
    data = f.readlines()

# vamos a convertir las lineas en registros
records = [eval(line.strip()) for line in data]

# Creamos el DataFrame a partir de los records o registros
df_UserItems= pd.DataFrame(records)

respaldo = df_UserItems

### 2nd Method for Uploading Raw Data: Gzip Format

# import gzip 
# import ast
# import pandas as pd
# data = []
# try:
#     with gzip.open('users_items.json.gz', 'rb') as d:
#         for line in d:
#             data.append (ast.literal_eval(line.decode('utf-8')))
#     Users_Items = pd.DataFrame (data)
# finally:
#     d.close()

## Explore and understand the dataset


# Quick view of the table
print('DataFrame - First & Last Rows:')
df_UserItems


## **Primary Analysis**
1. No missing values are observed in any column of the 'User_items' dataset.

2. The data collection process seems optimized as there are no missing values.

3. The 'items' column contains nested lists.

4. Further in-depth analysis will be performed.


# Check how many rows have null values in all columns
print(f'Total rows in df_UserItems: {df_UserItems.shape[0]}')
print(f'Total rows in df_UserItems with null values in all columns: {df_UserItems.isnull().all(axis=1).sum()}')


# General information about the DataFrame
print("General DataFrame Information:\n")
df_UserItems.info()


### **# Insights**
1. The **'items'** column contains a list with nested data (list of dictionaries).

2. We will normalize the column, resulting in a column for each key in the dictionary, maintaining traceability of the columns **('steam_id', 'items_count', 'user_id',** and **'user_url').**


# Normalize the 'items' column
df_UserItems = pd.json_normalize(records, record_path=['items'], meta=['steam_id', 'items_count', 'user_id', 'user_url'])


df_UserItems

# Find duplicate records
duplicates = df_UserItems[df_UserItems.duplicated(subset=['steam_id', 'item_id', 'playtime_forever'])]
duplicates.shape


# Remove duplicates
df_UserItems = df_UserItems.drop_duplicates(keep='first')


# Convert 'playtime_forever' column from minutes to hours
df_UserItems['hours_game'] = df_UserItems['playtime_forever'] / 60

# Round the values to two decimal places (optional)
df_UserItems['hours_game'] = df_UserItems['hours_game'].round(2)


# Get the top 5 maximum values
top_5_max = df_UserItems.nlargest(5, 'hours_game')

# Get the top 5 minimum values
top_5_min = df_UserItems.nsmallest(5, 'hours_game')

# Display the results
print("Top 5 maximum values:")
print(top_5_max[['hours_game']])

print("\nTop 5 minimum values:")
print(top_5_min[['hours_game']])


# Count the rows with a value of 0 in the 'hours_game' column
zero_rows = (df_UserItems['hours_game'] == 0).sum()

# Print the result
print(f"There are {zero_rows} rows with a value of 0 in the 'hours_game' column.")


# Remove rows with a value of 0 in the 'hours_game' column
df_UserItems = df_UserItems[df_UserItems['hours_game'] != 0]

# Remove irrelevant columns
df_UserItems = df_UserItems.drop(['playtime_2weeks','user_url','items_count','item_name','steam_id','playtime_forever'], axis=1)
df_UserItems.columns


## **# Insights:**
1. Remove rows with null values in this column for query optimization reasons.

2. There are 1,847,730 records with a value of 0 in the 'hours_game' column.

3. Irrelevant columns are also removed for query optimization reasons.


df_UserItems

df_UserItems.info()

## Clean data files are saved in the following formats for availability: CSV/Json/Parquet


# The files are stored locally
# df_UserItems.to_csv('user_items_clean.csv', index=False)
# df_UserItems.to_json('user_items_clean.json', orient='records', lines=True)
df_UserItems.to_parquet('../Data_Files/Cleaned_Data/2_user_items_clean.parquet', index=False)
