==> Jupyter_Notebooks/3_ETL_User_Reviews.ipynb <==
### **ETL process (Extract, Transform & Load) - Plataforma de videojuegos Stream**
#### **DATOS: User_Reviews**

Empezaremos por entender, transfomar y disponibilizar, encontrando problemas en los datos, aplicando tecnicas de limpieza y preprocesamiente. Finalmente, almacenaremos los datos transformatos para futuras exploraciones.  

Entenderemos el contexto y la informacion necesaria para nuestro analisis, tratando consistencia a esos datos.

**Nota: es necesario instalar las siguientes librerias dentro del entorno:**

*`pip install pandas numpy gdwon langdetect nltk` *




#### **Importamos las librerias necesarias**

import pandas as pd #para el analisis de tablas tabulares

import numpy as np # Para trabajar con arrays y matrices

import os # OS module ayuda a interactuar directamente con el sistema operativo

import gdown # facilita la descarga de archivos desde Google Drive utilizando su ID

import json # Python admite formato JSON para intercambiar datos de manera efectiva.

from langdetect import detect # Langdetect es una biblioteca para detectar automáticamente el idioma en el que está escrito un texto

import nltk # NLTK (Natural Language Toolkit) es una plataforma para construir programas Python para trabajar con datos de lenguaje humano

from nltk.sentiment import SentimentIntensityAnalyzer # SentimentIntensityAnalyzer es una herramienta en NLTK para análisis de sentimientos

'''
Al descargar 'vader_lexicon' puede devolver un error, verificación del certificado SSL al descargar el recurso desde el 
servidor NLTK. Porque el entorno de Python no puede verificar la validez del certificado SSL utilizado por el servidor NLTK. 
'''
import ssl # Por esto importamos SSL
ssl._create_default_https_context = ssl._create_unverified_context

nltk.download('vader_lexicon') # Nota: Este comando descarga el modelo de análisis de sentimiento de NLTK y solo debe ejecutarse la primera vez que se utiliza.


### **1. Función de Carga y Transformación del Archivo JSON Original**
Esta función realiza varios pasos para reformatear el archivo original, que tiene una estructura de diccionario de Python, en un formato JSON más adecuado. Aquí está lo que hace:

1. Verifica si el archivo original (en formato .json) ya está descargado. Si no lo está, lo descargará desde Google Drive, ya que el archivo no está en el formato correcto.

2. Aplica un formato adecuado al archivo para que quede un diccionario de Python dentro de un solo objeto JSON.

3. Guarda esos datos en un archivo llamado 'user_reviews_Estructurado.json'. y lo vamos a guardar como un data frame 'df_UserReviews'

def descargar_leer_json(url):
    """
    Descarga un archivo JSON desde la URL proporcionada, lo guarda localmente
    con el nombre 'datos_originales.json', lo convierte en registros y lo lee en un DataFrame de pandas.

    Parameters:
    - url (str): URL del archivo en Google Drive.

    Returns:
    - df (pd.DataFrame): DataFrame de pandas con los datos del archivo JSON.
    """

    # Descargar el archivo JSON si no existe localmente
    if not os.path.exists('user_reviews.json'):
        gdown.download(url, 'user_reviews.json', quiet=False)

    # Leer el archivo JSON descargado y convertirlo en texto
    with open('user_reviews.json', 'r', encoding='utf-8') as f:
        data = f.readlines()

    # Inicializar una lista vacía para almacenar los datos transformados
    transformed_data = []

    # Iterar sobre las líneas del archivo
    for line in data:
        # Eliminar los caracteres de nueva línea
        line = line.strip()
        # Convertir la línea en un diccionario de Python
        entry = eval(line)
        # Agregar la entrada transformada a la lista
        transformed_data.append(entry)

    # Escribir los datos transformados en un nuevo archivo JSON limpio
    with open('user_reviews_Estructurado.json', 'w') as json_file:
        json.dump(transformed_data, json_file, indent=4)

    # Leer el JSON transformado en un DataFrame de pandas
    df = pd.read_json('user_reviews_Estructurado.json')

    return df

### Llamamos la funcion para abrir el archivo

#URL del archivo en Google Drive y nombre del archivo local
url = 'https://drive.google.com/uc?export=download&id=1xQRf70-WS18UgjzAGKyJeTI30_f8wYyk'

#Utilizar la función (descargar_leer_json) para descargar y leer el JSON
df_UserReviews = descargar_leer_json(url)

# traemos una vista rapida de la tabla
df_UserReviews.head()

### **# Analisis Primario**
1. el Dataframe User_reviews esta compuesto de 3 columnas 'user_id', 'user_url', y 'reviews'

2. la columna **'reviews'** contiene informacion valiosa para nuestro estudio. Contiene datos en un diccionario estructurado en formato json e incluye informacion como reseñas y fecha de publicacion, entre otras.

3. vamos a dividir la informacion de La columna **'reviews'** en diferentes columnas para separarla para facilitar el analisis. 

### **2. Desanidamos los datos de la columna 'reviews' y los dividimos en diferentes columnas**

# Cargar el archivo user_reviews_Estructurado.json en un DataFrame
df_from_json = pd.read_json('user_reviews_Estructurado.json')

# Aplana la estructura JSON y crea un nuevo DataFrame
user_reviews_TF = pd.json_normalize(df_from_json.to_dict('records'), record_path=['reviews'], meta=['user_id', 'user_url'])

user_reviews_TF.head()

# Lista con el orden deseado de las columnas
column_order = ['user_id', 'user_url', 'funny', 'posted', 'last_edited', 'item_id', 'helpful', 'recommend', 'review']

# Reordenar las columnas del DataFrame
user_reviews_TF = user_reviews_TF[column_order]

# Ver los primeros registros del DataFrame resultante
user_reviews_TF.head()

# Obtener información general del DataFrame
print("\nInformación general del DataFrame:")
user_reviews_TF.info()

### **# Insights**
1. Las columnas tienen tipos de datos **'object'** y **'boolean'**.
   
2. la columna **'recommended'** es util para observar la distribucion de recomendaciones en este conjunto de datos

# Calcula valores en la columna 'recommend' 
conteo_por_recomendaciones = user_reviews_TF['recommend'].value_counts()
porcentaje_recomendaciones = user_reviews_TF['recommend'].value_counts(normalize=True) * 100
#porcentaje_recomendaciones = porcentaje_recomendaciones.round(2).astype(str) + '%'

# Crear un nuevo DataFrame con el conteo y porcentaje
resumen_recomendaciones = pd.DataFrame({
    'Conteo': conteo_por_recomendaciones,
    'Porcentaje': porcentaje_recomendaciones.round(2).astype(str) + '%'
})

# Ordenar el DataFrame por el conteo de mayor a menor
resumen_recomendaciones = resumen_recomendaciones.sort_values(by='Conteo', ascending=False)
resumen_recomendaciones

### **3. Limpieza Y preprocesamiento de los datos**

1. No hay necesidad de hacer procesos para eliminar Nulos, ya que el dataframe no cuenta con valores Nulos

2. Eliminamos columnas innecesarias

3. Realizaremos analisis de texto para determinar el idioma de las reseñas 

4. Categorizaremos las reseñas por medio de analisis de sentimiento 

5. Crearemos columnas derivadas de los datos

# Reemplazar valores vacíos, 'null' y 'None' con NaN en todo el DataFrame (No es necesario)
user_reviews_TF.replace(['', 'null', 'None'], np.nan, inplace=True)

# Filas donde todas las columnas especificadas tienen valores nulos.

columnas_con_nulos = user_reviews_TF.columns[user_reviews_TF.isnull().any()]
nulos_por_columna = user_reviews_TF[columnas_con_nulos].isnull().sum()

# Mostrar columnas con nulos y cantidad de nulos debajo de cada una
for columna in columnas_con_nulos:
    print(f"Columna: {columna}")
    print(f"Nulos: {nulos_por_columna[columna]}")

# Se establece umbral del 80% para decidir que columnas eliminar por valores nulos
Treshold_nulos = 0.8

# Calcula el porcentaje de valores nulos por columna
porcentaje_nulos = user_reviews_TF.isnull().mean()
  
# Filtra las columnas que superan el umbral
columnas_a_eliminar = porcentaje_nulos[porcentaje_nulos > Treshold_nulos]

# Muestra las columnas y su respectivo porcentaje de valores nulos
print("Columnas con más del {}% de valores nulos (candidatas a eliminar):".format(Treshold_nulos * 100))
for columna, porcentaje in columnas_a_eliminar.items():
    print("{}: {:.2%}".format(columna, porcentaje))

# Eliminamos las columnas que pasan el umbral establecido para valores nulos 
user_reviews_TF.drop(columns=columnas_a_eliminar.index, inplace=True)

# Se eliminan las columnas 'helpful' y 'user_url' por considerarse no relevantes
user_reviews_TF.drop(['helpful','user_url'], axis=1, inplace=True)

# Se buscan registros duplicados
user_reviews_TF.sort_values('user_id')
filas_duplicadas = user_reviews_TF[user_reviews_TF.duplicated(subset=['user_id', 'item_id', 'posted', 'review'], keep=False)]
filas_duplicadas.count()

# Se eliminan los registros duplicados basándome en múltiples columnas 
duplicados_eliminados = user_reviews_TF.drop_duplicates(subset=['user_id', 'item_id', 'posted', 'review'], keep=False, inplace=True)

cantidad_total_duplicados_eliminados = filas_duplicadas.shape[0]

print("Cantidad total de registros duplicados eliminados:", cantidad_total_duplicados_eliminados)

user_reviews_TF.head()

### **# Insights**
La columna **'posted'** contiene fechas de esas reviews, vamos a extraerlos 

1. vamosa  crear una nueva columna de tipo entero llamada **'year'**

2. vamos a extrar esa informacion de la columna **'posted'**

3. Se eliminara la columna **'posted'** por optimizacion
   
4. Tendremos un atributo en cada review con respecto al año, facilitando analisis de tendencias en el tiempo


# Visualizar los 5 primeros datos de la columna 'posted'
unique_posted_values = user_reviews_TF['posted'].head().unique()

for value in unique_posted_values:
    print(value)

### **# Insights**
La columna 'posted' contenia algunos valores faltantes en cuanto al año. con lo cual vamos a abordarlo de la siguiente manera

1. Utilizaremos la tecnica de interpolacion para llenar valores faltantes en la columna **'year'**, ya que este proceso es util cuando se tiene una relacion secuencial.

2.Elegimos utilizar **la interpolacion lineal por grupo** en cambio de aplicar **la interpolacion linear global**, ya que captura las posibles variaciones en la temporalidad entre diferentes juegos; tambien se alinea con la naturaleza de los datos, ya que la temporalidad varia entre juegos. 

3. Aplicaremos una agrupacion por juego, utilizando su identificador en la columna **'item_id'**  y asi tener una relacion de estas reseñas con relacion al tiempo con sus respectivos juegos

4. Aplicaremos **interpolacion lineal por grupo** en relacion a cada 'item_id'. Asi, estimaremos años faltantes usando la relacion con los años conocidos de el mismo juego.

5. Limpiaremos nuestro dataset de valores nulos despues de aplicar la interpolacion lineal por grupo, y llenaremos estos espacios con la mediana de la columna **'year'**

6. Con el paso anterior aseguramos que nuesto dataframe mantengauna coherencia de temporalidad necesaria para futuros analisis.

# Convertimos la columna 'posted' a tipo datetime
user_reviews_TF['posted'] = pd.to_datetime(user_reviews_TF['posted'].astype(str).str.replace(r'Posted |,|\.', '', regex=True), errors='coerce')

# Creamos la columna 'year' a partir de 'posted'
user_reviews_TF['year'] = user_reviews_TF['posted'].dt.year.astype('Int64')

# Ordena el DataFrame por 'item_id' y 'year' para asegurar que la interpolación se haga correctamente
user_reviews_TF = user_reviews_TF.sort_values(['item_id', 'year'])

# Rellenar valores nulos en 'year' mediante interpolación lineal por grupo (item_id)
user_reviews_TF['year'] = user_reviews_TF.groupby('item_id', group_keys=False)['year'].apply(lambda group: group.interpolate(method='pad') if group.notna().any() else group)

# Si aún hay valores nulos después de la interpolación, se llenan con la mediana. 
user_reviews_TF['year'] = user_reviews_TF['year'].fillna(user_reviews_TF['year'].median())

# Se eliminan las columnas 'posted' y 'user_id' que ya no son de utilidad
user_reviews_TF.drop(['posted', 'user_id'], axis=1, inplace=True)
user_reviews_TF.head()

### **3.1 Analisis de lenguaje**

Evaluando la columna 'review' podemos calcular la cantidad de reseñas por lenguage determinando:

1. El proceso de analisis de sentimiento se realiza en todo el conjunto de datos?, o

2. Es necesario aplicar un filtro especifico por idioma? 



# Análisis de texto para determinar el idioma de las reseñas

def detectar_idioma(texto):
    try:
        return detect(texto)
    except:
        return None


# Aplicar la función para detectar idioma y crear una nueva columna 'language'
user_reviews_TF['language'] = user_reviews_TF['review'].apply(detectar_idioma)


# Calcular el conteo y porcentaje de cada idioma
conteo_por_idioma = user_reviews_TF['language'].value_counts()
porcentaje_por_idioma = user_reviews_TF['language'].value_counts(normalize=True) * 100


# Crear un nuevo DataFrame con el conteo y porcentaje
resumen_idiomas = pd.DataFrame({
    'Conteo': conteo_por_idioma,
    'Porcentaje': porcentaje_por_idioma.round(2).astype(str) + '%'
})

# Ordenar el DataFrame por el conteo de mayor a menor
resumen_idiomas = resumen_idiomas.sort_values(by='Conteo', ascending=False)
resumen_idiomas.head()

# EDA information <<<--- 

# Mapeo de códigos de idioma a nombres completos (Top 5)
mapeo_idiomas = {
    'en': 'English',
    'pt': 'Portugués',
    'es': 'Español',
    'de': 'German',
    'so': 'Somali',
    # Agrega más mapeos según sea necesario
}

# Aplicar el mapeo al DataFrame
user_reviews_TF['idioma_completo'] = user_reviews_TF['language'].map(mapeo_idiomas)

# Crear un DataFrame con el resumen de idiomas
resumen_idiomas = user_reviews_TF['idioma_completo'].value_counts().reset_index()
resumen_idiomas.columns = ['Idioma', 'Conteo']
resumen_idiomas['Conteo'] = resumen_idiomas['Conteo'].round(2)

# Calcular el porcentaje
resumen_idiomas['Porcentaje'] = (resumen_idiomas['Conteo'] / len(user_reviews_TF)) * 100
resumen_idiomas['Porcentaje'] = resumen_idiomas['Porcentaje'].round(2)

# Guardar en un archivo CSV
resumen_idiomas.to_csv('../Dataframes/Dataframes_limpios/resumen_idiomas.csv', index=False)

resumen_idiomas

# Filtrar los registros donde language no es igual a 'en' 
user_reviews_TF = user_reviews_TF[user_reviews_TF['idioma_completo'] == 'English']

# Borrar la columna 'idioma_completo'
user_reviews_TF = user_reviews_TF.drop('idioma_completo', axis=1)

### **# Insights**
1. Utilizaremos los registros en ingles unicamente, ya que tenemos mas representacion, generando mas precision para nuestro modelo futuro y optimizacion de recursos computacionales. 

2. Al selececionar un metodo de analisis de sentimiento es necesario considerar factores como:
    - Precision del metodo
    - Facilidad de su implementacion 
    - eficiencia de recursos computacionales

3. Elegimos utilizar **la biblioteca NLTK (Natural Language Toolkit)** por los siguientes factores: 
   1. Facilidad en su implementacion 
   2. Eficiencia compuesta
   3. Polaridad compuesta
   4. Niveles aceptables de precision

### **Feature Engineering- NLP**

1. A partir de la columna **'review'** crearemos una columna sentiment_analysis usando **NLP** con los siguientes parametros:
   1. '0' = malo
   2. '1' = neutral
   3. '2' = positivo
   4. Al faltar una reseña asignaremos el valor de '1' por defecto

2. Eliminaremos las columnas de review despues de ejecutar el proceso de **NLP**, son innecesarias para el analisis. 

# Convertimos todas las letras a minúsculas para asegurar que todas las palabras sean tratadas de la misma manera.
user_reviews_TF.loc[:, 'review'] = user_reviews_TF['review'].str.lower()

#Eliminación de caracteres especiales
user_reviews_TF['review'] = user_reviews_TF['review'].replace('[^A-Za-z0-9\s]+', '', regex=True)

#Elimina caracteres de puntuación que no aportan al análisis de sentimiento.
user_reviews_TF.loc[:, 'review'] = user_reviews_TF['review'].str.replace('[^\w\s]', '', regex=True)

# Análisis de sentimiento para categorizar las reseñas columna 'review' 

def analyze_sentiments(df):
    # Instanciar el analizador de sentimientos
    sia = SentimentIntensityAnalyzer()

    # Aplicar el análisis de sentimientos y asignar valores numéricos
    df['compound_score'] = df['review'].apply(lambda review: sia.polarity_scores(review)['compound'])
    df['sentiment_analysis'] = df['compound_score'].apply(lambda score: 0 if score < 0 else (1 if score == 0 else 2))

    # Conteo de reviews por score
    score_counts = df['sentiment_analysis'].value_counts()

    # Conteo de reviews en blanco
    blank_reviews_count = df['review'].isnull().sum()

    # Total de reviews
    total_reviews = len(df)

    # Calcular porcentajes
    score_percentages = (score_counts / total_reviews * 100).round(2)
    blank_reviews_percentage = (blank_reviews_count / total_reviews * 100).round(2)
     
    # Se eliminan las columnas 'review' y 'compound_score', no necesitaremos estos datos
    df.drop(['review','compound_score'], axis=1, inplace=True)  

    return df, score_counts, blank_reviews_count, score_percentages, blank_reviews_percentage

# Llamar a la función analyze_sentiments
user_reviews_TF, score_counts, blank_reviews_count, score_percentages, blank_reviews_percentage = analyze_sentiments(user_reviews_TF)

# Crear un nuevo DataFrame con el conteo y porcentaje
resumen_sentimientos = pd.DataFrame({
    'Conteo': score_counts,
    'Porcentaje': score_percentages.round(2).astype(str) + '%'
})

# Ordenar el DataFrame por el conteo de mayor a menor
resumen_sentimientos = resumen_sentimientos.sort_values(by='Conteo', ascending=False)

# Imprimir los resultados
print("\nResumen de análisis de sentimientos:")
print(resumen_sentimientos)
print("\nConteo de reviews en blanco: ", blank_reviews_count, " Porcentaje: ", blank_reviews_percentage.round(2).astype(str) + '%')

# Borrar la columna 'language'
user_reviews_TF = user_reviews_TF.drop('language', axis=1)
user_reviews_TF

### 4. Guardamos los datos limpios
Guardamos estos archivos en los siguientes formatos para disponibilizarlos. CSV

# Los archivos se almacenan en local 
#user_reviews_TF.to_csv('user_reviews_limpios.csv', index=False)
#user_reviews_TF.to_json('user_reviews_limpios.json', orient='records', lines=True)
user_reviews_TF.to_parquet('../Dataframes/Dataframes_limpios/user_reviews_limpios.parquet', index=False)
