==> Jupyter_Notebooks/5_API_DF_.ipynb <==
## **# Creacion de Datasets para las APIS**

Vamos a disponibilizar, a la API, los datos limpios que dejaron como resultado nuestros procesos de ETL de los 3 datasets obtenidos de la plataforma **Steam**. 

        - 1_steam_games_limpios.parquet
        - 2_user_items_limpios.parquet
        - 3_user_reviews_limpios.parquet
   
El objetivo es la optimizacion de los recursos de computo y la eficiencia de los procesos.


### **Importamos las librerias** 

import pandas as pd # para el analisis de tablas tabulares

import os # permite interactuar con el sistema operativo

import ast # es útil cuando necesitas analizar o manipular código Python en un nivel más profundo que el proporcionado por el análisis de cadenas.

import numpy as np

steam_games_df = pd.read_parquet('../Data_Files/DF_limpios/1_steam_games_limpios.parquet')
user_items_df = pd.read_parquet('../Data_Files/DF_limpios/2_user_items_limpios.parquet')
user_reviews_df = pd.read_parquet('../Data_Files/DF_limpios/3_user_reviews_limpios.parquet')

steam_games_df.head(2)

user_items_df.head(2)

user_reviews_df.head(2)

## **Creacion de DF, especializados en cada Query**
#### nos traemos las columnas y las transformamos, con el objetivo de optimizar las consultas y solo enviar a la API los datos que son necesarios para cada proceso. 

.


### 1. def **PlayTimeGenre( genero : str )**: Debe devolver año con mas horas jugadas para dicho género.

Ejemplo de retorno: {"Año de lanzamiento con más horas jugadas para Género X" : 2013}



# Merge de los DataFrames
df_merged = pd.merge(user_items_df, steam_games_df[['item_id', 'year', 'genres']], on='item_id', how='inner')

# Explotar las listas en filas independientes
df_merged = df_merged.explode('genres').reset_index(drop=True)

# Agrupamos por año de lanzamiento y género y sumamos las horas jugadas
df_merged = df_merged.groupby(['year', 'genres'], as_index=False)['hours_game'].sum()

# Ordenar el DataFrame resultante
df_merged = df_merged.sort_values(by=['year', 'genres'])

# Guardar el resultado en un archivo CSV
df_merged.to_parquet('../Data_Files/API_Files/PlayTimeGenre.parquet', index=False)


df_merged

###  2. def **UserForGenre( genero : str )**: Debe devolver el usuario que acumula más horas jugadas para el género dado y una lista de la acumulación de horas jugadas por año.

Ejemplo de retorno: {"Usuario con más horas jugadas para Género X" : us213ndjss09sdf, "Horas jugadas":[{Año: 2013, Horas: 203}, {Año: 2012, Horas: 100}, {Año: 2011, Horas: 23}]}

df_game = steam_games_df.drop(['name', 'developer', 'price', 'tags','specs'],axis=1)

# Merge de los DataFrames
merged_data=pd.merge(user_items_df,df_game,left_on='item_id', right_on='item_id', how='right')

#borramos item_id, no es relevante
merged_data.drop(['item_id'], axis=1, inplace=True)

#borramos nulos
merged_data.dropna(inplace=True)

#cambiamos el tipo de dato de hours game a int-- nos da las horas. Igual con Year
merged_data['hours_game'] = merged_data['hours_game'].astype(int)
merged_data['year'] = merged_data['year'].astype(int)

#Convertimos la columna genres a tipo lista con el modulo ast.
#merged_data['genres'] = merged_data['genres'].apply(ast.literal_eval)

#Expandimos las listas en columnas
merged_data = merged_data.explode('genres').reset_index(drop=True)

#Encontrar el jugador con más horas jugadas por género y año
max_hours_df = merged_data.loc[merged_data.groupby(['genres', 'year'])['hours_game'].idxmax()]

#Damos un orden a las columnas 
max_hours_df = max_hours_df[['genres', 'year', 'user_id', 'hours_game']]

#Segunda agrupación para obtener las horas jugadas por año para el usuario seleccionado
max_hours = max_hours_df.groupby(['genres', 'year', 'user_id'])['hours_game'].sum().reset_index()

max_hours.to_parquet('../Data_Files/API_Files/UserForGenre.parquet', index=False)


merged_data

max_hours_df

###  3. **def UsersRecommend( año : int )**: Devuelve el top 3 de juegos MÁS recomendados por usuarios para el año dado. (reviews.recommend = True y comentarios positivos/neutrales)

Ejemplo de retorno: [{"Puesto 1" : X}, {"Puesto 2" : Y},{"Puesto 3" : Z}]

# Merge de los DataFrames
df_merged = pd.merge(user_reviews_df, steam_games_df, on='item_id', how='left')

# Filtrar las filas que cumplen con las condiciones
result = df_merged.loc[(df_merged['recommend'] == True) & (df_merged['sentiment_analysis'].isin([1, 2])),
                           ['year_x', 'name']]

# Renombrar la columna 'year_x' a 'year'
result = result.rename(columns={'year_x': 'year'})

# Agrupar por 'year' y 'name' y contar las ocurrencias
grouped_result = result.groupby(['year', 'name']).size().reset_index(name='count')

# Ordenar por 'year' y 'count' en orden descendente
grouped_result = grouped_result.sort_values(by=['year', 'count'], ascending=[False, False])

# Obtener el top 3 por año
top3_by_year = grouped_result.groupby('year').head(3)

# Guardar el resultado en un archivo CSV
top3_by_year.to_parquet('../Data_Files/API_Files/UsersRecommend.parquet', index=False)

result

top3_by_year

grouped_result

top3_by_year

### 4. def **UsersNotRecommend( año : int )**: Devuelve el top 3 de juegos MENOS recomendados por usuarios para el año dado. (reviews.recommend = False y comentarios negativos)
   
Ejemplo de retorno: [{"Puesto 1" : X}, {"Puesto 2" : Y},{"Puesto 3" : Z}]

# Merge de los DataFrames
df_merged = pd.merge(user_reviews_df, steam_games_df, on='item_id', how='left')

# Filtrar las filas que cumplen con las condiciones
result = df_merged.loc[(df_merged['recommend'] == True) & (df_merged['sentiment_analysis'].isin([0])),
                           ['year_x', 'name']]

# Renombrar la columna 'year_x' a 'year'
result = result.rename(columns={'year_x': 'year'})

# Agrupar por 'year' y 'name' y contar las ocurrencias
grouped_result = result.groupby(['year', 'name']).size().reset_index(name='count')

# Ordenar por 'year' y 'count' en orden descendente
grouped_result = grouped_result.sort_values(by=['year', 'count'], ascending=[False, False])

# Obtener el top 3 por año menos recomendados
top3_least_recommended_by_year = grouped_result.groupby('year').tail(3)

# Guardar el resultado en un archivo CSV
top3_least_recommended_by_year.to_parquet('../Data_Files/API_Files/UsersNotRecommend.parquet', index=False)

result

grouped_result

top3_least_recommended_by_year

### 5. def **sentiment_analysis( año : int )**: Según el año de lanzamiento, se devuelve una lista con la cantidad de registros de reseñas de usuarios que se encuentren categorizados con un análisis de sentimiento.

Ejemplo de retorno: {Negative = 182, Neutral = 120, Positive = 278}

Steam_modif= steam_games_df.copy()

Steam_modif.rename(columns={'year': 'year_released'}, inplace=True)

# Merge de los DataFrames
merged_df = pd.merge(Steam_modif, user_reviews_df, on='item_id', how='inner')

# Agrupar por año_released y sentimiento, y contar las ocurrencias
result_df = merged_df.groupby(['year_released', 'sentiment_analysis']).size().reset_index(name='count')

# Pivoteando el DataFrame para tener sentimientos como columnas
result_df = result_df.pivot(index='year_released', columns='sentiment_analysis', values='count').reset_index()

# Renombrar las columnas según el formato deseado
result_df.columns = ['year_released', 'Negative', 'Neutral', 'Positive']

# Rellenar NaN con 0
result_df = result_df.fillna(0)

# Convertir las columnas a tipo int
result_df[['Negative', 'Neutral', 'Positive']] = result_df[['Negative', 'Neutral', 'Positive']].astype(int)

# Guardar el resultado en un archivo CSV
result_df.to_parquet('../Data_Files/API_Files/sentiment_analysis.parquet', index=False)

result_df
